ref #265 "added suppor validation cards before remove data"

+++ src/GUI/private/mainmodel.cpp
@@ -595,8 +595,43 @@ void MainModel::handleCardEditFinished(const QSharedPointer<API::Card>& card) {
 
 void MainModel::handleRemoveRequest(const QSharedPointer<API::Card> &card) {
 
-    _currentCardsListModel->removeCard(card->cardId());
-    _db->deleteObject(card);
+
+    auto service = QmlNotificationService::NotificationService::getService();
+
+    if (service) {
+
+        QmlNotificationService::Listner listner = [card, this] (bool accepted) {
+
+            if (accepted) {
+
+                _currentCardsListModel->removeCard(card->cardId());
+                _db->deleteObject(card);
+            }
+        };
+
+        if (getMode()) {
+            auto listOfUsers = _backEndModel->getAllUserFromCard(card->cardId());
+
+            if (listOfUsers.size()) {
+                service->setNotify(tr("Operation not permitted"),
+                                tr("This card have a active clients, so you can't to remove this card."),
+                                "",
+                                QmlNotificationService::NotificationData::Error);
+                return;
+            }
+
+            listner(true);
+            return;
+        }
+
+
+        service->setQuestion(listner, tr("Remove Card"),
+                             tr("You trying to delete this card, do not worry a seller that has give out this card save all bonuses locally, "
+                                " so after repeat visit you will be get all your removed bonuses again."
+                                " Do you want to continue?"));
+
+
+    }
 }
 
 void MainModel::handleCardSelectedForWork(const QSharedPointer<CardModel> &card) {
